requirements.txt
google-adk==1.14.1
google-genai==1.38.0
numpy==2.0.2
pandas==2.3.2
